v32 Phase A Step 2 完了 — tests 全緑、v32 feature flag 追加
- Cargo.toml: v32 = ["v27"] feature flag 新設
- tests: u64 eid 化に伴うキャスト(as u32 / as u64)、HashSet<u64> 化、tie → tie_ref 切替
- src/engine.rs, src/ravn.rs: インラインテストの u64 対応
- 229 tests pass (lib 110 + integration 119)
- examples はまだコンパイル不通(次で対処)

+++ examples/agentic_workload_bench.rs
@@ -36,9 +36,9 @@ fn main() {
     eng.define_himo("decision", HimoType::Symbol, 0);
     eng.flush().unwrap();
 
-    let speaker_eids: Vec<u32> = (0..N_SPEAKERS).map(|_| eng.entity()).collect();
-    let topic_eids: Vec<u32> = (0..N_TOPICS).map(|_| eng.entity()).collect();
-    let file_eids: Vec<u32> = (0..N_FILES).map(|_| eng.entity()).collect();
+    let speaker_eids: Vec<u64> = (0..N_SPEAKERS).map(|_| eng.entity()).collect();
+    let topic_eids: Vec<u64> = (0..N_TOPICS).map(|_| eng.entity()).collect();
+    let file_eids: Vec<u64> = (0..N_FILES).map(|_| eng.entity()).collect();
 
     println!("── INSERT ──");
     let t0 = Instant::now();
